+++ Runtime/Auth/AuthClientResult.cs
@@ -14,5 +14,6 @@ namespace Auth
         public string UserName { get; set; }
         public string Email { get; set; }
         public string AccessToken { get; set; }
+        public string RefreshToken { get; set; }
     }
 }