+++ Runtime/Data/YoureUser.cs
@@ -14,7 +14,8 @@ namespace Data
         public string Email { get; set; }
         public string UserName { get; set; }
         public string AccessToken { get; set; }
-       // public bool NewsletterAccepted{ get; set; }
-       // public bool TermsAccepted{ get; set; }
+
+        // public bool NewsletterAccepted{ get; set; }
+        // public bool TermsAccepted{ get; set; }
     }
 }